+++ src/plugins/switch/cray_aries/util.c
@@ -71,8 +71,8 @@ int create_apid_dir(uint64_t apid, uid_t uid, gid_t gid)
 
 	rc = chown(apid_dir, uid, gid);
 	if (rc) {
-		CRAY_ERR("chown %s, %d, %d failed: %m",
-			 apid_dir, (int)uid, (int)gid);
+		CRAY_ERR("chown %s, %u, %u failed: %m",
+			 apid_dir, uid, gid);
 		xfree(apid_dir);
 		return SLURM_ERROR;
 	}
